% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunc.R
\name{fars_read}
\alias{fars_read}
\title{Read in File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A user specific file path of the .csv file to be imported into R environment}
}
\value{
This function creates a data frame of the subclass tibble from the .csv file path provided.
    A readout will be provided in the console with information on the rows and columns included.
}
\description{
Simple function to read in a csv file by path and produce data table read-out in console
    with information on number of rows and column header names.
}
\note{
Error will appear if no file path name is provided or if file path does not exist on user's computer.
}
