% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuncs.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Generates a table of months and years from fars data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{integer}
}
\value{
A list of Months and Years.
}
\description{
Function to generate a table of months available for a list of years.
}
\details{
Will produce errors if data file for a particular year is not available.
}
\examples{
fars_read_years()
fars_read_years(2017)
fars_read_years(c(2013,2014,2015))

}
