% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuncs.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generates a table showing number of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{integer}
}
\value{
A table showing total number of accident by month and year.
}
\description{
Function to generate a table showing numbers of accidents for a list of years from fars data.
}
\examples{
fars_summarize_years()
fars_summarize_years(2017)
fars_summarize_years(c(2013,2014,2015))

}
