% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunc.R
\name{make_filename}
\alias{make_filename}
\title{Create file Name based on year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year the user would like printed in the file name tag.}
}
\value{
This function will print the file name path with the year specified in the function.
}
\description{
This function prints to the console the inputted year within a common file name path.
}
\examples{
make_filename(2015)

}
