% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{map FARS data by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num, }{state from which the data will be displayed (number 1- 56)}

\item{year}{, year from which the data will be displayed}
}
\value{
This function returns a graphical object.
}
\description{
map accidents by state and year with US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS)
}
\details{
it is necessary to download the fars files before, this functions read the files from the working directory, if not an error will be thrown
Throws an error of "invalid State Number:" for invalid state number.
}
\examples{
library(dplyr)
library(readr)
library(tidyr)
library(maps)
library(graphics)
f13path<-system.file("extdata", "accident_2013.csv.bz2", package = "farsfuncs")
file.copy(from=c(f13path),to=getwd())
fars_map_state(1, 2013)

}
