% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{read FARS files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{, name of file to be read}
}
\value{
a data.frame with the data of file
}
\description{
read FARS data (US National Highway Traffic Safety Administration's Fatality Analysis Reporting System)from file
}
\details{
it is necessary to download the fars files before, this functions read the files from the working directory, if not an error will be thrown
Throws an error if wrong file name is provided
}
