% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunc.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create Month & Year Tibble}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The year of the data the user would like be used in the function.}
}
\value{
A summary of the tibble will be printed to the console, with information on the two columns and number of rows.
}
\description{
This function creates a tibble with a 'MONTH' column and a 'year' column.
}
\note{
If a year is entered into the function that cannot be located on the user's computer,
         an error message will appear.
}
