% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{read FARS files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{, a list with all years to be displayed}
}
\value{
a data.frame with the month and the total of accidents by year
}
\description{
read FARS data from file
}
\details{
it is necessary to download the fars files before, this functions read the files from the working directory, if not a warning
Throws a warning of "invalid year:"
   if a numeric other than what is available in data is provided as year.
}
\examples{
library(dplyr)
library(readr)
library(tidyr)
fars_read_years(c(2013))

}
