% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunc.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create Number of Rows by Month in Specified Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The year of the data the user would like be used in the function.}
}
\value{
A summary of the tibble will be printed to the console, with one column indicating months 1 (January) -
    12 (December), and a second column, titled with 'years' parameter. The value in the second column is the number of rows
    found for that month in that year within the dataset.
}
\description{
This function creates a tibble with a 'MONTH' column and a 'year' column that returns the
  number of rows for each month (1-12) in the year inputted to the function.
}
\note{
If a year is entered into the function that cannot be connected to a dataset located on the user's computer,
         an error message will appear.
}
