% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{total FARS accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{, a list with all years to be displayed}
}
\value{
a data.frame with the month and the total of accidents by year
}
\description{
read FARS data from file and count the number of accidents group_by year and month
}
\details{
it is necessary to download the fars files before, this functions read the files from the working directory, if not a warning will be thrown
Throws a warning of "invalid year:" if a numeric other than
  what is available is provided as years
}
\examples{
library(dplyr)
library(readr)
library(tidyr)
fars_summarize_years(c(2013))

}
